init config data only for deeptest demo site

+++ internal/server/modules/source/config.go
@@ -1,6 +1,7 @@
 package source
 
 import (
+	"github.com/aaronchen2k/deeptest/internal/pkg/config"
 	"github.com/aaronchen2k/deeptest/internal/server/modules/model"
 	"github.com/aaronchen2k/deeptest/internal/server/modules/repo"
 	"github.com/gookit/color"
@@ -10,19 +11,23 @@ type ConfigSource struct {
 	ConfigRepo *repo.ConfigRepo `inject:""`
 }
 
-func (s *ConfigSource) GetSources() ([]model.SysConfig, error) {
-	configs := []model.SysConfig{
-		{
-			Key:   "agentUrlOpts",
-			Value: "[     {         \"label\": \"本地\",         \"value\": \"http://127.0.0.1:8086/api/v1\",         \"desc\": \"客户端默认在本机启动的代理\"     },     {         \"label\": \"演示站点\",         \"value\": \"http://111.231.16.35:8086/api/v1\",         \"desc\": \"DeepTest测试站点代理\"     }  ]",
-		},
+func (s *ConfigSource) GetSources() (configs []model.SysConfig, err error) {
+	if config.CONFIG.System.SysEnv == "deeptest_demo" {
+		configs = []model.SysConfig{
+			{
+				Key:   "agentUrlOpts",
+				Value: "[     {         \"label\": \"本地\",         \"value\": \"http://127.0.0.1:8086/api/v1\",         \"desc\": \"客户端默认在本机启动的代理\"     },     {         \"label\": \"演示站点\",         \"value\": \"http://111.231.16.35:8086/api/v1\",         \"desc\": \"DeepTest测试站点代理\"     }  ]",
+			},
+		}
 	}
 
-	return configs, nil
+	return
 }
 
 func (s *ConfigSource) Init() error {
-	s.ConfigRepo.DB.Delete(&model.SysConfig{}, "1=1")
+	if config.CONFIG.System.SysEnv == "deeptest_demo" {
+		s.ConfigRepo.DB.Delete(&model.SysConfig{}, "1=1")
+	}
 
 	sources, err := s.GetSources()
 	if err != nil {
